:white_check_mark: Fix many tests for new annotations

+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/expression/function/scalar/math/Acos.java
@@ -22,7 +22,7 @@ import java.util.List;
  */
 public class Acos extends AbstractTrigonometricFunction {
     @FunctionInfo(returnType = "double", description = "The arccosine of an angle, expressed in radians.")
-    public Acos(Source source, @Param(name = "n", type = { "integer", "long", "double", "unsigned_long" }) Expression n) {
+    public Acos(Source source, @Param(name = "n", type = { "double", "integer", "long", "unsigned_long" }) Expression n) {
         super(source, n);
     }
